removed recipe mod in goose server and moved it into router mod
