--- crates/goose-server/src/recipe/mod.rs
@@ -1,2 +1,0 @@
-pub mod list_recipes;
-pub mod recipe_manifest_metadata;
